+++ e2e/call-canister.spec.ts
@@ -3,12 +3,12 @@ import {initTestSuite} from './utils/init.utils';
 
 const getPartyPage = initTestSuite();
 
-testWithII('should call canister to request an approval of some ICP', async () => {
+testWithII('should call canister to transfer some ICP', async () => {
   const partyPage = getPartyPage();
 
   await partyPage.approvePermissionsAccounts();
 
-  await partyPage.approvePermissionsConsentMessageIcrc2Approve();
+  await partyPage.approvePermissionsConsentMessage();
 
-  await partyPage.icrc2Approve();
+  await partyPage.callCanister();
 });
